requirements.txt
mss==9.0.1
numpy==2.0.0
opencv-python==4.10.0.84
overlay==1.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1